requirements.txt
requests==2.21.0
Django==2.1.7
beautifulsoup4==4.7.1
djangorestframework==3.9.4
coreapi==2.3.3
gunicorn
whitenoise
django-environ
psycopg2-binary
flake8
django-import-export
django-nose
coverage
sentry-sdk
django-lockdown
drf-yasg
pytest
pytest-cov
pytest-django